+++ src/iframe-helper.js
@@ -588,7 +588,7 @@ export function canInspectWindow(win) {
 }
 
 /** @const {string} */
-export const EMBED_PROP = '__AMP_EMBED__';
+export const FIE_EMBED_PROP = '__AMP_EMBED__';
 
 /**
  * Returns the embed created using `installFriendlyIframeEmbed` or `null`.
@@ -600,6 +600,17 @@ export const EMBED_PROP = '__AMP_EMBED__';
  */
 export function getFriendlyIframeEmbedOptional(iframe) {
   return /** @type {?./friendly-iframe-embed.FriendlyIframeEmbed} */ (iframe[
-    EMBED_PROP
+    FIE_EMBED_PROP
   ]);
+}
+
+/**
+ * @param {!Element} element
+ * @return {boolean}
+ */
+export function isInFie(element) {
+  return (
+    element.classList.contains('i-amphtml-fie') ||
+    !!closestAncestorElementBySelector(element, '.i-amphtml-fie')
+  );
 }
