+++ src/service/layers-impl.js
@@ -18,7 +18,8 @@ import {Services} from '../services';
 import {computedStyle} from '../style';
 import {dev, devAssert} from '../log';
 import {getMode} from '../mode';
-import {isConnectedNode, isInFie, rootNodeFor} from '../dom';
+import {isConnectedNode, rootNodeFor} from '../dom';
+import {isInFie} from '../iframe-helper';
 import {listen} from '../event-helper';
 import {registerServiceBuilderForDoc} from '../service';
 import {remove} from '../utils/array';
